refactor(sensing): move uriFilter into the sensing layer
