+++ extension/test/unit/services/sensing/uriFilter.test.ts
@@ -17,7 +17,7 @@
 import * as vscode from 'vscode';
 import * as assert from 'assert';
 
-import { shouldRecordUri, shouldRecordUriString } from '@extension/services/telemetry/uriFilter';
+import { shouldRecordUri, shouldRecordUriString } from '@extension/services/sensing/uriFilter';
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
